Fix: Loginメソッドをmoduleとして切り出した
